refactor(installer): split at.py god file into paths, tui, and CLI dispatcher
Extract the shared install-root constants into installer/paths.py and the
entire TUI concern (tab rendering, the reconcile loop, Esc handling) into
installer/tui.py, leaving at.py as the CLI/entry dispatcher. Behavior is
unchanged; tests are retargeted for the new import/monkeypatch locations only.
